feat : UserCash 잔돈 계산 메서드 변수 선언

+++ src/main/java/domain/ChangeCalculator.java
@@ -4,12 +4,12 @@ import java.util.EnumMap;
 
 public class ChangeCalculator {
 
-    public EnumMap<Bill, Integer> calculateChange(int balance) {
+    public EnumMap<Bill, Integer> calculateChange(long balance) {
         EnumMap<Bill, Integer> result = new EnumMap<>(Bill.class);
 
         for (Bill coin : Bill.values()) {
             int coinValue = coin.getValue();
-            int count = balance / coinValue;  // 각 동전의 개수를 계산
+            int count = (int) balance / coinValue;  // 각 동전의 개수를 계산
             balance %= coinValue;             // 잔액을 업데이트
             result.put(coin, count);
         }
